stock-model/requirements.txt
absl-py==0.9.0
alembic==1.4.1
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
configparser==4.0.2
databricks-cli==0.10.0
docker==4.2.0
entrypoints==0.3
Flask==1.1.2
gast==0.2.2
gitdb==4.0.2
GitPython==3.1.0
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gorilla==0.3.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
mlflow==1.7.2
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==0.25.3
pkg-resources==0.0.0
prometheus-client==0.7.1
prometheus-flask-exporter==0.13.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
sklearn==0.0
smmap==3.0.1
SQLAlchemy==1.3.13
sqlparse==0.3.1
tabulate==0.8.7
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
